New modal logic
Used "createPortal" for modal window.
Modal component can reused now.
Divided ContactAddForm.
