+++ src/components/Contacts/ContactAdd/ContactFormLabel/ContactFormLabel.module.scss
@@ -2,12 +2,15 @@
 
 .label {
   @include flex;
-  font-size: 22px;
+  justify-content: center;
   width: 100%;
   margin-bottom: 15px;
 
   font-size: 12px;
   color: red;
+  & label:last-child {
+    margin-bottom: 30px;
+  }
   &Name {
     font-size: 12px;
     color: #171717;
