+++ src/components/Contacts/modules/interfaces.ts
@@ -1,0 +1,4 @@
+export interface MyFormValues {
+  name: string;
+  number: string;
+}
