+++ src/shared/components/Modal/Modal.module.scss
@@ -1,0 +1,69 @@
+@import "../../../index.scss";
+
+.overlay {
+  top: 0;
+  bottom: 0;
+  left: 0;
+  right: 0;
+
+  width: 100%;
+  height: 100%;
+
+  position: absolute;
+  z-index: 999;
+
+  background-color: rgba(0, 0, 0, 0.3);
+
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.modal {
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  width: 80%;
+  height: 60%;
+  background-color: white;
+
+  box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
+
+  border-radius: 15px;
+  @include tablet {
+    width: 50%;
+    height: 60%;
+  }
+}
+
+.closeButton {
+  position: absolute;
+  top: 2%;
+  right: 2%;
+  width: 20px;
+  height: 20px;
+  border-style: none;
+  border-radius: 50%;
+  padding: 0;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  cursor: pointer;
+
+  transition-property: transform;
+  transition-duration: 250ms;
+
+  &:hover,
+  &:focus {
+    transform: rotate(20deg);
+  }
+
+  @include tablet {
+    width: 40px;
+    height: 40px;
+    font-size: 28px;
+  }
+}
